refactor: 웹과 콘솔 어플리케이션 메인 클래스를 각 패키지로 분리
