+++ src/main/java/roomescape/console/RoomescapeConsoleApplication.java
@@ -1,9 +1,11 @@
-package roomescape;
+package roomescape.console;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = {"roomescape.console", "roomescape.core"})
 public class RoomescapeConsoleApplication {
     public static void main(String[] args) {
         SpringApplication.run(RoomescapeConsoleApplication.class, args);
