+++ src/main/java/roomescape/web/RoomescapeWebApplication.java
@@ -1,8 +1,10 @@
-package roomescape;
+package roomescape.web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
+@ComponentScan(basePackages = {"roomescape.web", "roomescape.core"})
 @SpringBootApplication
 public class RoomescapeWebApplication {
     public static void main(String[] args) {
